Switching to coloredlogs from clint
clint does not have an easy way to level log messages
and filter them based on user preference. This is useful
to Surgen as surgen is very verbose with it's output, and
oftentimes it is difficult to analyze the results of a larger
set of targets in a single go.

Standardizing around logging gives consumers of Surgen
the ability to use the log levels they are familiar with
to programatically control Surgen output.

It also has the added benefit of allowing one to customize
the log formatter and handler if used programatically.