requiring python3.6, adding utilities
due to the EOL of python at the end of this year, and
the fact that surgen is intended to be a command line tool,
we can be aggressive a timeline and only support newer versions
of Python.

The value of 3.5 and above is type annotations, which would help
heavily in terms of documentation.