using short summary for VCS comment, enabling envvars in git
the summary becomes really long with large (10+)
number of surgen procedures, and the skip
is not relevant to the person consuming the change.

As a result, changing the summary to use a
short summary of only successful changes instead.

Introducing a way to pass through environment
variables to the git target as a commit is
being made, enabling things like setting GIT_COMMITTER_NAME.